+++ src/main/java/corecord/dev/domain/user/service/UserService.java
@@ -3,7 +3,7 @@ package corecord.dev.domain.user.service;
 import corecord.dev.common.exception.GeneralException;
 import corecord.dev.common.status.ErrorStatus;
 import corecord.dev.common.util.CookieUtil;
-import corecord.dev.common.util.JwtUtil;
+import corecord.dev.domain.auth.util.JwtUtil;
 import corecord.dev.domain.analysis.repository.AbilityRepository;
 import corecord.dev.domain.analysis.repository.AnalysisRepository;
 import corecord.dev.domain.folder.repository.FolderRepository;
